Update node-websocket-server to v.1.3.00 + event proxying patch
This version of node-websocket-server has an unofficial patch that makes
event proxying from http.Server to ws.Server work just like in v.1.1.00
when server hijacking is not used. This is because how event proxying will
work in the future (if it will be supported at all) is currently undecided.